+++ Producto/Back-end/kydira-api/kydira-api/src/main/java/com/kydira_api/model/User.java
@@ -15,7 +15,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userId;
+    private Long userId;
 
     private String firstName;
